refactor: structural foundation under structs (host-aligned seams)
- Expression dual ownership: expressionType/valueType/ValueForm
- TypeQuery productCanAssignFrom + arraySubscriptInfo
- symbols::AnnotationStore on AST; AddressPlan (Field/Index); CallPlan
- SA/CSNB splits; resolveMemberBase trusts ValueForm

No brace initializers or function-pointer surface.
