+++ src/ast/AbstractSyntaxTree.h
@@ -7,12 +7,14 @@
 
 #include "ast/AbstractSyntaxTreeNode.h"
 #include "parser/SyntaxTree.h"
+#include "symbols/AnnotationStore.h"
 
 namespace ast {
 
 class AbstractSyntaxTree: public parser::SyntaxTree {
 private:
     std::vector<std::unique_ptr<AbstractSyntaxTreeNode>> translationUnit;
+    symbols::AnnotationStore annotations_;
 
 public:
     AbstractSyntaxTree(std::vector<std::unique_ptr<AbstractSyntaxTreeNode>> translationUnit);
@@ -20,6 +22,9 @@ public:
 
     auto begin() const -> decltype(translationUnit.begin());
     auto end() const -> decltype(translationUnit.end());
+
+    symbols::AnnotationStore& annotations() { return annotations_; }
+    const symbols::AnnotationStore& annotations() const { return annotations_; }
 
     void accept(ast::AbstractSyntaxTreeVisitor& visitor) const;
     void accept(parser::SyntaxTreeVisitor& visitor) override;
