+++ src/ast/CMakeLists.txt
@@ -13,6 +13,7 @@ add_library(ast
         Constant.cpp
         ConstantExpression.cpp
         ContextualSyntaxNodeBuilder.cpp
+        CSNB_Creators.cpp
         Declaration.cpp
         DeclarationSpecifiers.cpp
         Declarator.cpp
@@ -130,4 +131,4 @@ add_library(ast
     )
 
 trans_target_defaults(ast)
-target_link_libraries(ast PUBLIC parser types semantic_analyzer)
+target_link_libraries(ast PUBLIC parser types semantic_analyzer symbols)
