+++ src/ast/Expression.cpp
@@ -8,16 +8,35 @@ void Expression::setType(const type::Type& type) {
     this->type = type;
 }
 
-type::Type Expression::getType() const {
+type::Type Expression::expressionType() const {
     if (!type) {
-        throw std::runtime_error { "type is not set" };
+        throw std::runtime_error { "expression type is not set" };
     }
     return *type;
 }
 
-void Expression::setResultSymbol(semantic_analyzer::ValueEntry resultSymbol) {
-    this->resultSymbol = std::make_unique<semantic_analyzer::ValueEntry>(resultSymbol);
+type::Type Expression::valueType() const {
+    if (resultSymbol) {
+        return resultSymbol->getType();
+    }
+    return expressionType();
+}
+
+bool Expression::hasDecayedArrayValue() const {
+    return isArrayObjectType() && hasResultSymbol() && valueType().isPointer();
+}
+
+void Expression::setTypeAndResult(semantic_analyzer::ValueEntry resultSymbol) {
+    this->resultSymbol = std::make_unique<semantic_analyzer::ValueEntry>(std::move(resultSymbol));
     setType(this->resultSymbol->getType());
+    form = ValueForm::Scalar;
+}
+
+void Expression::setAggregateAddressResult(semantic_analyzer::ValueEntry addressSymbol,
+        const type::Type& aggregateType) {
+    this->resultSymbol = std::make_unique<semantic_analyzer::ValueEntry>(std::move(addressSymbol));
+    setType(aggregateType);
+    form = ValueForm::AggregateAddress;
 }
 
 bool Expression::hasResultSymbol() const {
@@ -38,4 +57,3 @@ semantic_analyzer::ValueEntry* Expression::getLvalueSymbol() const {
 }
 
 } // namespace ast
-
