+++ src/codegen/Amd64Registers.h
@@ -21,6 +21,8 @@ public:
     Register& getMultiplicationRegister();
     Register& getRemainderRegister();
     Register& getCounterRegister();
+    // Caller-saved, non-arg register for indirect call targets (r10).
+    Register& getIndirectCallTargetRegister();
 
 private:
     Register rax { "rax" };
