+++ src/codegen/AssemblyGenerator.cpp
@@ -87,7 +87,11 @@ void AssemblyGenerator::generateCodeFor(const Argument& argument) {
 }
 
 void codegen::AssemblyGenerator::generateCodeFor(const Call& call) {
-    stackMachine->callProcedure(call.getProcedureName());
+    if (call.isIndirect()) {
+        stackMachine->callProcedureIndirect(call.getProcedureName());
+    } else {
+        stackMachine->callProcedure(call.getProcedureName());
+    }
 }
 
 void codegen::AssemblyGenerator::generateCodeFor(const Return& returnCommand) {
