+++ src/codegen/quadruples/Call.h
@@ -9,17 +9,21 @@ namespace codegen {
 
 class Call: public Quadruple {
 public:
-    Call(std::string procedureName);
+    // Direct call by procedure label, or indirect when `indirect` is true
+    // (procedureName is then the symbol holding the callee address).
+    Call(std::string procedureName, bool indirect = false);
     virtual ~Call() = default;
 
     void generateCode(AssemblyGenerator& generator) const override;
 
     std::string getProcedureName() const;
+    bool isIndirect() const;
 
 private:
     void print(std::ostream& stream) const override;
 
     std::string procedureName;
+    bool indirect { false };
 };
 
 } // namespace codegen
