+++ src/semantic_analyzer/CMakeLists.txt
@@ -2,6 +2,9 @@ add_library(semantic_analyzer
         FunctionEntry.cpp
         LabelEntry.cpp
         SemanticAnalysisVisitor.cpp
+        SemanticAnalysisVisitor_Calls.cpp
+        SemanticAnalysisVisitor_Expressions.cpp
+        SemanticAnalysisVisitor_Statements.cpp
         SemanticAnalyzer.cpp
         SemanticXmlOutputVisitor.cpp
         SymbolTable.cpp
@@ -10,6 +13,7 @@ add_library(semantic_analyzer
         FunctionEntry.h
         LabelEntry.h
         SemanticAnalysisVisitor.h
+        SemanticAnalysisVisitorInternal.h
         SemanticAnalyzer.h
         SemanticXmlOutputVisitor.h
         SymbolTable.h
@@ -17,4 +21,4 @@ add_library(semantic_analyzer
         ValueScope.h)
 
 trans_target_defaults(semantic_analyzer)
-target_link_libraries(semantic_analyzer PUBLIC translation_unit types ast)
+target_link_libraries(semantic_analyzer PUBLIC translation_unit types ast symbols)
