+++ src/semantic_analyzer/SemanticAnalysisVisitor.cpp
@@ -1,41 +1,13 @@
-#include "SemanticAnalysisVisitor.h"
+#include "SemanticAnalysisVisitorInternal.h"
 
-#include <algorithm>
-#include <cctype>
-#include <limits>
-#include <stdexcept>
-#include <string>
-
-#include "translation_unit/Context.h"
-#include "types/Type.h"
-#include "util/Logger.h"
-#include "util/LogManager.h"
+#include "ast/IdentifierExpression.h"
 
 namespace semantic_analyzer {
 
-static const translation_unit::Context EXTERNAL_CONTEXT {"external", 0};
-
-static Logger& err = LogManager::getErrorLogger();
-
-// Locals are stored as `$s<scopeId><name>` in the symbol table; strip that prefix for diagnostics
-// and for looking up file-scope functions under their source names.
-static std::string unscopedSymbolName(const std::string& name) {
-    if (name.size() > 2 && name[0] == '$' && name[1] == 's') {
-        std::size_t i = 2;
-        while (i < name.size() && std::isdigit(static_cast<unsigned char>(name[i]))) {
-            ++i;
-        }
-        if (i > 2 && i < name.size()) {
-            return name.substr(i);
-        }
-    }
-    return name;
-}
-
 SemanticAnalysisVisitor::SemanticAnalysisVisitor() {
     type::Type functionType = type::function(type::signedInteger());
-    symbolTable.insertFunction("printf", functionType.getFunction(), EXTERNAL_CONTEXT);
-    symbolTable.insertFunction("scanf", functionType.getFunction(), EXTERNAL_CONTEXT);
+    symbolTable.insertFunction("printf", functionType.getFunction(), externalContext());
+    symbolTable.insertFunction("scanf", functionType.getFunction(), externalContext());
 }
 
 SemanticAnalysisVisitor::~SemanticAnalysisVisitor() {
@@ -78,7 +50,6 @@ void SemanticAnalysisVisitor::visit(ast::Declaration& declaration) {
                     declarator->getContext());
         } else if (symbolTable.insertSymbol(declarator->getName(), type, declarator->getContext())) {
             declarator->setHolder(symbolTable.lookup(declarator->getName()));
-            // TODO: type check initializers
             if (declarator->hasInitializer() && symbolTable.isAtFileScope()) {
                 long initValue = 0;
                 if (declarator->getInitializer()->evaluateConstant(initValue)) {
@@ -103,732 +74,6 @@ void SemanticAnalysisVisitor::visit(ast::Declarator& declarator) {
 
 void SemanticAnalysisVisitor::visit(ast::InitializedDeclarator& declarator) {
     declarator.visitChildren(*this);
-}
-
-void SemanticAnalysisVisitor::visit(ast::ArrayAccess& arrayAccess) {
-    arrayAccess.visitLeftOperand(*this);
-    arrayAccess.visitRightOperand(*this);
-
-    // Operand failed to resolve (e.g. undeclared identifier) — error already reported.
-    if (!arrayAccess.hasLeftOperandSymbol() || !arrayAccess.hasRightOperandSymbol()) {
-        return;
-    }
-
-    auto type = arrayAccess.leftOperandType();
-    const type::Type leftSymType = arrayAccess.leftOperandSymbol()->getType();
-    type::Type elementType = type::voidType();
-    int stride = 0;
-    bool baseIsArray = false;
-    // Multi-dim: a[i] may keep expression type T[N] while the value is already a decayed pointer.
-    if (type.isArray() && leftSymType.isPointer()) {
-        elementType = type.getElementType();
-        stride = elementType.getSize();
-        baseIsArray = false;
-    } else if (type.isArray()) {
-        elementType = type.getElementType();
-        stride = elementType.getSize();
-        baseIsArray = true;
-    } else if (type.isPointer() || leftSymType.isPointer()) {
-        type::Type ptr = type.isPointer() ? type : leftSymType;
-        elementType = ptr.dereference();
-        stride = elementType.getSize();
-        baseIsArray = false;
-    } else {
-        semanticError("invalid type for operator[]\n", arrayAccess.getContext());
-        return;
-    }
-    if (stride <= 0) {
-        stride = 1;
-    }
-    arrayAccess.setBaseIsArray(baseIsArray);
-    arrayAccess.setElementSize(stride);
-    if (elementType.isArray() || elementType.isStructure()) {
-        // a[i] for multi-dim or array-of-struct: yield address of the element so
-        // a[i].member stores into the array, not a loaded temp.
-        type::Type addrType = elementType.isArray()
-                ? type::pointer(elementType.getElementType())
-                : type::pointer(elementType);
-        auto addr = symbolTable.createTemporarySymbol(addrType);
-        arrayAccess.setLvalue(addr);
-        arrayAccess.setResultSymbol(addr);
-        arrayAccess.setType(elementType);
-        arrayAccess.setYieldsAddress(true);
-    } else {
-        // Lvalue holds element address; result is a loaded rvalue.
-        arrayAccess.setLvalue(symbolTable.createTemporarySymbol(type::pointer(elementType)));
-        arrayAccess.setResultSymbol(symbolTable.createTemporarySymbol(elementType));
-        arrayAccess.setType(elementType);
-        arrayAccess.setYieldsAddress(false);
-    }
-}
-
-void SemanticAnalysisVisitor::visit(ast::MemberAccess& memberAccess) {
-    memberAccess.getBase()->accept(*this);
-    if (!memberAccess.getBase()->getResultSymbol()) {
-        return;
-    }
-    type::Type baseType = memberAccess.getBase()->getType();
-    type::Type valueType = memberAccess.getBase()->getResultSymbol()->getType();
-    type::Type structureType = type::voidType();
-    bool baseIsPointer = false;
-
-    if (memberAccess.isArrow()) {
-        // -> requires expression type pointer-to-structure (not dual-type nested aggregates).
-        if (baseType.isPointer()) {
-            structureType = baseType.dereference();
-            baseIsPointer = true;
-        } else if (valueType.isPointer() && !baseType.isStructure()) {
-            // Declared pointer whose expression type was not updated.
-            structureType = valueType.dereference();
-            baseIsPointer = true;
-        } else {
-            semanticError("base of ‘->’ is not a pointer to structure", memberAccess.getContext());
-            return;
-        }
-        if (!structureType.isStructure()) {
-            semanticError("base of ‘->’ is not a pointer to structure", memberAccess.getContext());
-            return;
-        }
-    } else {
-        // Dot: structure expression type, or dual-type nested aggregate (value already address).
-        if (baseType.isStructure()) {
-            structureType = baseType;
-            // Nested aggregate dual-type: value is pointer-to-struct while expr type is struct.
-            baseIsPointer = valueType.isPointer();
-        } else if (valueType.isStructure() && !valueType.isPointer()) {
-            structureType = valueType;
-            baseIsPointer = false;
-        } else if (baseType.isPointer() || valueType.isPointer()) {
-            // `p.x` on a pointer is invalid C (must use ->).
-            semanticError("request for member in non-structure type", memberAccess.getContext());
-            return;
-        } else {
-            semanticError("request for member in non-structure type", memberAccess.getContext());
-            return;
-        }
-    }
-
-    int offset = 0;
-    type::Type memberType = type::voidType();
-    if (!structureType.memberOffset(memberAccess.getMemberName(), offset)
-            || !structureType.memberType(memberAccess.getMemberName(), memberType)) {
-        semanticError("no member named ‘" + memberAccess.getMemberName() + "’ in structure",
-                memberAccess.getContext());
-        return;
-    }
-    memberAccess.setMemberOffset(offset);
-    memberAccess.setBaseIsPointer(baseIsPointer);
-    memberAccess.setFieldAddressSymbol(symbolTable.createTemporarySymbol(type::pointer(memberType)));
-    if (memberType.isStructure() || memberType.isArray()) {
-        // Nested aggregate: result is address (dual-type like multi-dim rows).
-        memberAccess.setResultSymbol(*memberAccess.getFieldAddressSymbol());
-        memberAccess.setType(memberType);
-    } else {
-        memberAccess.setResultSymbol(symbolTable.createTemporarySymbol(memberType));
-        memberAccess.setType(memberType);
-    }
-}
-
-void SemanticAnalysisVisitor::visit(ast::FunctionCall& functionCall) {
-    functionCall.visitOperand(*this);
-    functionCall.visitArguments(*this);
-
-    // Operand failed to resolve (e.g. undeclared identifier) — error already reported.
-    if (!functionCall.hasOperandSymbol()) {
-        return;
-    }
-
-    // ValueEntry names for locals are scope-prefixed (e.g. `$s1a`); functions are stored under
-    // the source identifier. A mangled name is always a local (or temp), never a function entry —
-    // do not look up the demangled form, or a local would incorrectly call a same-named function.
-    const auto symbolName = functionCall.operandSymbol()->getName();
-    const auto displayName = unscopedSymbolName(symbolName);
-    if (symbolName != displayName || !symbolTable.hasFunction(displayName)) {
-        semanticError("called object `" + displayName + "` is not a function", functionCall.getContext());
-        return;
-    }
-
-    auto functionSymbol = symbolTable.findFunction(displayName);
-
-    functionCall.setSymbol(functionSymbol);
-
-    auto& arguments = functionCall.getArgumentList();
-    // Reject function designators as values (e.g. printf("%d", main)) — codegen has no address.
-    for (auto& argument : arguments) {
-        if (argument->hasResultSymbol()) {
-            rejectFunctionValue(argument->getResultSymbol()->getType(), functionCall.getContext());
-        }
-    }
-
-    if (arguments.size() == functionSymbol.argumentCount()) {
-        auto declaredArguments = functionSymbol.arguments();
-        for (std::size_t i { 0 }; i < arguments.size(); ++i) {
-            if (!arguments.at(i)->hasResultSymbol()) {
-                return;
-            }
-            const auto& declaredArgument = declaredArguments.at(i);
-            const auto& actualArgument = arguments.at(i)->getResultSymbol();
-            typeCheck(actualArgument->getType(), declaredArgument, functionCall.getContext());
-        }
-
-        auto returnType = functionSymbol.returnType();
-        if (!returnType.isVoid()) {
-            functionCall.setResultSymbol(symbolTable.createTemporarySymbol(returnType));
-        }
-    } else if (functionSymbol.getContext() == EXTERNAL_CONTEXT) {
-    // FIXME: using EXTERNAL_CONTEXT as a workaround for printf/scanf external functions until varargs are properly implemented
-        auto returnType = functionSymbol.returnType();
-        if (!returnType.isVoid()) {
-            functionCall.setResultSymbol(symbolTable.createTemporarySymbol(returnType));
-        }
-    } else {
-        semanticError("no match for function " + functionSymbol.getType().to_string(), functionCall.getContext());
-    }
-}
-
-void SemanticAnalysisVisitor::visit(ast::IdentifierExpression& identifier) {
-    if (symbolTable.hasSymbol(identifier.getIdentifier())) {
-        identifier.setResultSymbol(symbolTable.lookup(identifier.getIdentifier()));
-    } else {
-        semanticError("symbol `" + identifier.getIdentifier() + "` is not defined", identifier.getContext());
-    }
-}
-
-void SemanticAnalysisVisitor::visit(ast::ConstantExpression& constant) {
-    constant.setResultSymbol(symbolTable.createTemporarySymbol(constant.getType()));
-}
-
-void SemanticAnalysisVisitor::visit(ast::StringLiteralExpression& stringLiteral) {
-    std::string constantSymbol = symbolTable.newConstant(stringLiteral.getValue());
-    stringLiteral.setConstantSymbol(constantSymbol);
-    stringLiteral.setResultSymbol(symbolTable.createTemporarySymbol(stringLiteral.getType()));
-}
-
-void SemanticAnalysisVisitor::visit(ast::PostfixExpression& expression) {
-    expression.visitOperand(*this);
-    if (!expression.hasOperandSymbol()) {
-        return;
-    }
-    rejectFunctionValue(expression.operandType(), expression.getContext());
-
-    expression.setType(expression.operandType());
-    auto operandSymbol = *expression.operandSymbol();
-    expression.setResultSymbol(operandSymbol);
-
-    auto preOperationSymbolName = operandSymbol.getName() + "_pre";
-    symbolTable.insertSymbol(preOperationSymbolName, operandSymbol.getType(), operandSymbol.getContext());
-    expression.setPreOperationSymbol(symbolTable.lookup(preOperationSymbolName));
-
-    if (!expression.isLval()) {
-        semanticError("lvalue required as increment operand", expression.getContext());
-    }
-}
-
-void SemanticAnalysisVisitor::visit(ast::PrefixExpression& expression) {
-    expression.visitOperand(*this);
-    if (!expression.hasOperandSymbol()) {
-        return;
-    }
-    rejectFunctionValue(expression.operandType(), expression.getContext());
-
-    expression.setType(expression.operandType());
-    expression.setResultSymbol(*expression.operandSymbol());
-
-    if (!expression.isLval()) {
-        semanticError("lvalue required as increment operand", expression.getContext());
-    }
-}
-
-void SemanticAnalysisVisitor::visit(ast::UnaryExpression& expression) {
-    const auto& lexeme = expression.getOperator()->getLexeme();
-    if (lexeme == "sizeof") {
-        expression.visitOperand(*this);
-        if (!expression.hasOperandSymbol()) {
-            expression.setResultSymbol(symbolTable.createTemporarySymbol(type::signedInteger()));
-            return;
-        }
-        const type::Type& operandType = expression.operandType();
-        // Mirror sizeof(type): void and bare function types are incomplete for sizeof.
-        // Pointers (including pointer-to-function) remain complete; those types also
-        // report isFunction() because pointer() copies the function payload.
-        if (operandType.isVoid() || (operandType.isFunction() && !operandType.isPointer())) {
-            semanticError(
-                    "invalid application of ‘sizeof’ to incomplete type ‘" + operandType.to_string() + "’",
-                    expression.getContext());
-            expression.setResultSymbol(symbolTable.createTemporarySymbol(type::signedInteger()));
-            return;
-        }
-        expression.setSizeofValue(operandType.getSize());
-        expression.setResultSymbol(symbolTable.createTemporarySymbol(type::signedInteger()));
-        return;
-    }
-
-    expression.visitOperand(*this);
-    if (!expression.hasOperandSymbol()) {
-        return;
-    }
-    rejectFunctionValue(expression.operandType(), expression.getContext());
-
-    switch (lexeme.front()) {
-    case '&':
-        expression.setResultSymbol(symbolTable.createTemporarySymbol(type::pointer(expression.operandType())));
-        break;
-    case '*': {
-        type::Type operandType = expression.operandType();
-        const type::Type valueType = expression.operandSymbol()->getType();
-        // Value already a pointer (e.g. multi-dim a[i] decayed row, or int(*)[N]).
-        if (valueType.isPointer()) {
-            type::Type pointee = valueType.dereference();
-            if (pointee.isArray()) {
-                // *ptr-to-array yields the array object (address); do not scalar-load the row.
-                auto addr = symbolTable.createTemporarySymbol(type::pointer(pointee.getElementType()));
-                expression.setLvalueSymbol(addr);
-                expression.setResultSymbol(addr);
-                expression.setType(pointee);
-            } else {
-                expression.setResultSymbol(symbolTable.createTemporarySymbol(pointee));
-                expression.setLvalueSymbol(symbolTable.createTemporarySymbol(valueType));
-            }
-            break;
-        }
-        // Array object in memory: *a ≡ a[0].
-        if (operandType.isArray()) {
-            type::Type elem = operandType.getElementType();
-            if (elem.isArray()) {
-                // *a for multi-dim: yield decayed address of first row; keep array expr type.
-                auto addr = symbolTable.createTemporarySymbol(type::pointer(elem.getElementType()));
-                expression.setLvalueSymbol(addr);
-                expression.setResultSymbol(addr);
-                expression.setType(elem);
-            } else {
-                auto addr = symbolTable.createTemporarySymbol(type::pointer(elem));
-                expression.setLvalueSymbol(addr);
-                expression.setResultSymbol(symbolTable.createTemporarySymbol(elem));
-                expression.setType(elem);
-            }
-            break;
-        }
-        semanticError("invalid type argument of ‘unary *’ :" + operandType.to_string(), expression.getContext());
-        break;
-    }
-    case '+':
-        expression.setResultSymbol(*expression.operandSymbol());
-        break;
-    case '-':
-        expression.setResultSymbol(symbolTable.createTemporarySymbol(expression.operandType()));
-        break;
-    case '~':
-        expression.setResultSymbol(symbolTable.createTemporarySymbol(expression.operandType()));
-        break;
-    case '!':
-        expression.setResultSymbol(symbolTable.createTemporarySymbol(type::signedInteger()));
-        expression.setTruthyLabel(symbolTable.newLabel());
-        expression.setFalsyLabel(symbolTable.newLabel());
-        break;
-    default:
-        throw std::runtime_error { "Unidentified unary operator: " + expression.getOperator()->getLexeme() };
-    }
-}
-
-void SemanticAnalysisVisitor::visit(ast::TypeCast& expression) {
-    expression.visitOperand(*this);
-    if (!expression.hasOperandSymbol()) {
-        return;
-    }
-
-    type::Type target = expression.getTypeSpecifier().getType();
-    if (target.isArray() || (target.isFunction() && !target.isPointer())) {
-        semanticError("cast to array or function type ‘" + target.to_string() + "’", expression.getContext());
-        return;
-    }
-
-    type::Type source = expression.operandType();
-    if (source.isFunction() && !source.isPointer()) {
-        semanticError("cast of function designator is not supported", expression.getContext());
-        return;
-    }
-    // Operand may be an array object or a dual-type multi-dim row (value already a pointer).
-    // Codegen materializes AddressOf only when the value type is still an array.
-    expression.setResultSymbol(symbolTable.createTemporarySymbol(target));
-}
-
-void SemanticAnalysisVisitor::visit(ast::ArithmeticExpression& expression) {
-    expression.visitLeftOperand(*this);
-    expression.visitRightOperand(*this);
-    if (!expression.hasLeftOperandSymbol() || !expression.hasRightOperandSymbol()) {
-        return;
-    }
-    rejectFunctionValue(expression.leftOperandType(), expression.getContext());
-    rejectFunctionValue(expression.rightOperandType(), expression.getContext());
-
-    typeCheck(
-            expression.leftOperandType(),
-            expression.rightOperandType(),
-            expression.getContext());
-    // FIXME: type conversion
-    expression.setResultSymbol(symbolTable.createTemporarySymbol(expression.leftOperandType()));
-}
-
-void SemanticAnalysisVisitor::visit(ast::ShiftExpression& expression) {
-    expression.visitLeftOperand(*this);
-    expression.visitRightOperand(*this);
-    if (!expression.hasLeftOperandSymbol() || !expression.hasRightOperandSymbol()) {
-        return;
-    }
-    rejectFunctionValue(expression.leftOperandType(), expression.getContext());
-    rejectFunctionValue(expression.rightOperandType(), expression.getContext());
-
-    if (expression.rightOperandType().isPrimitive() && !expression.rightOperandType().getPrimitive().isFloating()) {
-        expression.setResultSymbol(symbolTable.createTemporarySymbol(expression.leftOperandType()));
-    } else {
-        semanticError("argument of type int required for shift expression", expression.getContext());
-    }
-}
-
-void SemanticAnalysisVisitor::visit(ast::ComparisonExpression& expression) {
-    expression.visitLeftOperand(*this);
-    expression.visitRightOperand(*this);
-    if (!expression.hasLeftOperandSymbol() || !expression.hasRightOperandSymbol()) {
-        return;
-    }
-    rejectFunctionValue(expression.leftOperandType(), expression.getContext());
-    rejectFunctionValue(expression.rightOperandType(), expression.getContext());
-
-    typeCheck(
-            expression.leftOperandType(),
-            expression.rightOperandType(),
-            expression.getContext());
-
-    expression.setResultSymbol(symbolTable.createTemporarySymbol(type::signedInteger()));
-    expression.setTruthyLabel(symbolTable.newLabel());
-    expression.setFalsyLabel(symbolTable.newLabel());
-}
-
-void SemanticAnalysisVisitor::visit(ast::BitwiseExpression& expression) {
-    expression.visitLeftOperand(*this);
-    expression.visitRightOperand(*this);
-    if (!expression.hasLeftOperandSymbol() || !expression.hasRightOperandSymbol()) {
-        return;
-    }
-    rejectFunctionValue(expression.leftOperandType(), expression.getContext());
-    rejectFunctionValue(expression.rightOperandType(), expression.getContext());
-    expression.setType(expression.leftOperandType());
-
-    typeCheck(
-            expression.leftOperandType(),
-            expression.rightOperandType(),
-            expression.getContext());
-
-    expression.setResultSymbol(
-            symbolTable.createTemporarySymbol(expression.getType()));
-}
-
-void SemanticAnalysisVisitor::visit(ast::LogicalAndExpression& expression) {
-    expression.visitLeftOperand(*this);
-    expression.visitRightOperand(*this);
-    if (!expression.hasLeftOperandSymbol() || !expression.hasRightOperandSymbol()) {
-        return;
-    }
-    rejectFunctionValue(expression.leftOperandType(), expression.getContext());
-    rejectFunctionValue(expression.rightOperandType(), expression.getContext());
-
-    typeCheck(
-            expression.leftOperandType(),
-            expression.rightOperandType(),
-            expression.getContext());
-
-    expression.setResultSymbol(symbolTable.createTemporarySymbol(type::signedInteger()));
-    expression.setExitLabel(symbolTable.newLabel());
-}
-
-void SemanticAnalysisVisitor::visit(ast::LogicalOrExpression& expression) {
-    expression.visitLeftOperand(*this);
-    expression.visitRightOperand(*this);
-    if (!expression.hasLeftOperandSymbol() || !expression.hasRightOperandSymbol()) {
-        return;
-    }
-    rejectFunctionValue(expression.leftOperandType(), expression.getContext());
-    rejectFunctionValue(expression.rightOperandType(), expression.getContext());
-
-    typeCheck(
-            expression.leftOperandType(),
-            expression.rightOperandType(),
-            expression.getContext());
-
-    expression.setResultSymbol(symbolTable.createTemporarySymbol(type::signedInteger()));
-    expression.setExitLabel(symbolTable.newLabel());
-}
-
-void SemanticAnalysisVisitor::visit(ast::ConditionalExpression& expression) {
-    expression.visitCondition(*this);
-    expression.visitTrueExpression(*this);
-    expression.visitFalseExpression(*this);
-
-    if (!expression.getCondition()->hasResultSymbol()
-            || !expression.getTrueExpression()->hasResultSymbol()
-            || !expression.getFalseExpression()->hasResultSymbol()) {
-        return;
-    }
-
-    rejectFunctionValue(expression.conditionSymbol()->getType(), expression.getContext());
-    rejectFunctionValue(expression.trueSymbol()->getType(), expression.getContext());
-    rejectFunctionValue(expression.falseSymbol()->getType(), expression.getContext());
-
-    typeCheck(
-            expression.trueSymbol()->getType(),
-            expression.falseSymbol()->getType(),
-            expression.getContext());
-
-    // Result type follows the true arm after typeCheck (same policy as other binary ops).
-    const type::Type resultType = expression.trueSymbol()->getType();
-    expression.setType(resultType);
-    expression.setResultSymbol(symbolTable.createTemporarySymbol(resultType));
-    expression.setFalsyLabel(symbolTable.newLabel());
-    expression.setExitLabel(symbolTable.newLabel());
-}
-
-void SemanticAnalysisVisitor::visit(ast::AssignmentExpression& expression) {
-    expression.visitLeftOperand(*this);
-    expression.visitRightOperand(*this);
-    if (!expression.hasLeftOperandSymbol() || !expression.hasRightOperandSymbol()) {
-        return;
-    }
-
-    if (expression.isLval()) {
-        rejectFunctionValue(expression.leftOperandType(), expression.getContext());
-        rejectFunctionValue(expression.rightOperandType(), expression.getContext());
-        // Dual-type aggregates (nested member / a[i] as struct) hold a pointer-sized
-        // value while expression type is structure — whole-struct Assign would copy
-        // the address, not the object. Reject until aggregate copy is implemented.
-        const type::Type rightExpr = expression.rightOperandType();
-        const type::Type rightValue = expression.rightOperandSymbol()->getType();
-        if (rightExpr.isStructure() && rightValue.isPointer()) {
-            semanticError("assignment of structure from dual-type aggregate is not supported",
-                    expression.getContext());
-            return;
-        }
-        typeCheck(
-                expression.leftOperandType(),
-                expression.rightOperandType(),
-                expression.getContext());
-
-        expression.setResultSymbol(*expression.leftOperandSymbol());
-    } else {
-        semanticError("lvalue required on the left side of assignment", expression.getContext());
-    }
-}
-
-void SemanticAnalysisVisitor::visit(ast::ExpressionList& expression) {
-    expression.visitLeftOperand(*this);
-    expression.visitRightOperand(*this);
-    if (!expression.hasRightOperandSymbol()) {
-        return;
-    }
-    // Comma operator: value and type of the right operand
-    expression.setType(expression.rightOperandType());
-    expression.setResultSymbol(*expression.rightOperandSymbol());
-}
-
-void SemanticAnalysisVisitor::visit(ast::Operator&) {
-}
-
-void SemanticAnalysisVisitor::visit(ast::JumpStatement& statement) {
-    if (loopStack.empty()) {
-        semanticError("`" + statement.jumpKeyword.type + "` statement not in loop or switch",
-                statement.jumpKeyword.context);
-        return;
-    }
-    const auto& loop = loopStack.back();
-    if (statement.jumpKeyword.type == "break") {
-        statement.setJumpTo(*loop.exit);
-    } else if (statement.jumpKeyword.type == "continue") {
-        if (!loop.cont) {
-            semanticError("`continue` statement not in loop", statement.jumpKeyword.context);
-            return;
-        }
-        statement.setJumpTo(*loop.cont);
-    } else {
-        semanticError("unsupported jump statement `" + statement.jumpKeyword.type + "`", statement.jumpKeyword.context);
-    }
-}
-
-void SemanticAnalysisVisitor::visit(ast::SwitchStatement& statement) {
-    statement.expression->accept(*this);
-    if (statement.expression->hasResultSymbol()) {
-        rejectFunctionValue(statement.expression->getResultSymbol()->getType(),
-                statement.expression->getContext());
-    }
-
-    auto exitLabel = symbolTable.newLabel();
-    statement.setExitLabel(exitLabel);
-    statement.setCaseTemp(symbolTable.createTemporarySymbol(type::signedInteger()));
-
-    LabelEntry* continueLabel = nullptr;
-    if (!loopStack.empty()) {
-        continueLabel = loopStack.back().cont;
-    }
-    // break → switch exit; continue only if nested in a loop (cont may be null).
-    loopStack.push_back({ nullptr, continueLabel, statement.getExitLabel() });
-    switchStack.push_back(&statement);
-
-    statement.body->accept(*this);
-
-    switchStack.pop_back();
-    loopStack.pop_back();
-}
-
-void SemanticAnalysisVisitor::visit(ast::CaseLabel& statement) {
-    // Always attach a codegen label so the node is well-formed even when illegal.
-    statement.setLabel(symbolTable.newLabel());
-
-    if (switchStack.empty()) {
-        semanticError("case label not within a switch statement", statement.caseExpression->getContext());
-        statement.statement->accept(*this);
-        return;
-    }
-
-    statement.caseExpression->accept(*this);
-    long value = 0;
-    if (!statement.caseExpression->evaluateConstant(value)) {
-        semanticError("case label is not a constant expression", statement.caseExpression->getContext());
-        statement.statement->accept(*this);
-        return;
-    }
-    statement.setCaseValue(value);
-    for (const auto* existing : switchStack.back()->getCases()) {
-        if (existing->getCaseValue() == value) {
-            semanticError("duplicate case value", statement.caseExpression->getContext());
-            statement.statement->accept(*this);
-            return;
-        }
-    }
-    switchStack.back()->addCase(&statement);
-
-    statement.statement->accept(*this);
-}
-
-void SemanticAnalysisVisitor::visit(ast::DefaultLabel& statement) {
-    // Always attach a label for codegen, even when the label is illegal / duplicate.
-    statement.setLabel(symbolTable.newLabel());
-
-    if (switchStack.empty()) {
-        semanticError("default label not within a switch statement", statement.defaultKeyword.context);
-        statement.statement->accept(*this);
-        return;
-    }
-
-    if (switchStack.back()->getDefaultLabel()) {
-        // Keep the first default for codegen; ignore subsequent ones as targets.
-        semanticError("multiple default labels in switch", statement.defaultKeyword.context);
-    } else {
-        switchStack.back()->setDefaultLabel(&statement);
-    }
-
-    statement.statement->accept(*this);
-}
-
-void SemanticAnalysisVisitor::visit(ast::GotoStatement& statement) {
-    pendingGotos.push_back(&statement);
-}
-
-void SemanticAnalysisVisitor::visit(ast::LabeledStatement& statement) {
-    // Always attach a codegen label so the statement node is well-formed even when
-    // the name is a duplicate (goto targets keep the first definition only).
-    auto label = symbolTable.newLabel();
-    statement.setLabel(label);
-    if (namedLabels.find(statement.getLabelName()) != namedLabels.end()) {
-        semanticError("duplicate label `" + statement.getLabelName() + "`", statement.name.context);
-    } else {
-        namedLabels.insert({ statement.getLabelName(), label });
-    }
-    statement.statement->accept(*this);
-}
-
-void SemanticAnalysisVisitor::visit(ast::ReturnStatement& statement) {
-    statement.returnExpression->accept(*this);
-    if (statement.returnExpression->hasResultSymbol()) {
-        rejectFunctionValue(statement.returnExpression->getResultSymbol()->getType(),
-                statement.returnExpression->getContext());
-    }
-}
-
-void SemanticAnalysisVisitor::visit(ast::VoidReturnStatement& statement) {
-}
-
-void SemanticAnalysisVisitor::visit(ast::IfStatement& statement) {
-    statement.testExpression->accept(*this);
-    if (statement.testExpression->hasResultSymbol()) {
-        rejectFunctionValue(statement.testExpression->getResultSymbol()->getType(),
-                statement.testExpression->getContext());
-    }
-    statement.body->accept(*this);
-
-    statement.setFalsyLabel(symbolTable.newLabel());
-}
-
-void SemanticAnalysisVisitor::visit(ast::IfElseStatement& statement) {
-    statement.testExpression->accept(*this);
-    if (statement.testExpression->hasResultSymbol()) {
-        rejectFunctionValue(statement.testExpression->getResultSymbol()->getType(),
-                statement.testExpression->getContext());
-    }
-    statement.truthyBody->accept(*this);
-    statement.falsyBody->accept(*this);
-
-    statement.setFalsyLabel(symbolTable.newLabel());
-    statement.setExitLabel(symbolTable.newLabel());
-}
-
-void SemanticAnalysisVisitor::visit(ast::LoopStatement& loop) {
-    const bool declScope = loop.header->opensBlockScope();
-    if (declScope) {
-        symbolTable.enterBlockScope();
-    }
-    loop.header->accept(*this);
-    // for-with-increment: continue before increment. while: continue → entry.
-    // do-while: header preassigns continue (before the test); leave it alone.
-    if (loop.header->increment) {
-        loop.header->setLoopContinue(symbolTable.newLabel());
-    } else if (loop.header->continueTargetsEntry()) {
-        loop.header->setLoopContinue(*loop.header->getLoopEntry());
-    }
-    loopStack.push_back({ loop.header->getLoopEntry(), loop.header->getLoopContinue(), loop.header->getLoopExit() });
-    loop.body->accept(*this);
-    loopStack.pop_back();
-    if (declScope) {
-        symbolTable.exitBlockScope();
-    }
-}
-
-void SemanticAnalysisVisitor::visit(ast::ForLoopHeader& loopHeader) {
-    if (loopHeader.initialization) {
-        loopHeader.initialization->accept(*this);
-    }
-    if (loopHeader.clause) {
-        loopHeader.clause->accept(*this);
-    }
-    if (loopHeader.increment) {
-        loopHeader.increment->accept(*this);
-    }
-
-    loopHeader.setLoopEntry(symbolTable.newLabel());
-    loopHeader.setLoopExit(symbolTable.newLabel());
-}
-
-void SemanticAnalysisVisitor::visit(ast::WhileLoopHeader& loopHeader) {
-    loopHeader.clause->accept(*this);
-
-    loopHeader.setLoopEntry(symbolTable.newLabel());
-    loopHeader.setLoopExit(symbolTable.newLabel());
-}
-
-void SemanticAnalysisVisitor::visit(ast::DoWhileLoopHeader& loopHeader) {
-    loopHeader.clause->accept(*this);
-
-    loopHeader.setLoopEntry(symbolTable.newLabel());
-    // continue jumps here (re-test), not to the body entry.
-    loopHeader.setLoopContinue(symbolTable.newLabel());
-    loopHeader.setLoopExit(symbolTable.newLabel());
 }
 
 void SemanticAnalysisVisitor::visit(ast::Pointer&) {
@@ -914,6 +159,7 @@ void SemanticAnalysisVisitor::visit(ast::FunctionDefinition& function) {
         return;
     }
     function.setSymbol(symbolTable.findFunction(function.getName()));
+    currentReturnType = function.getSymbol()->returnType();
     symbolTable.startFunction(function.getName(), argumentNames);
     namedLabels.clear();
     pendingGotos.clear();
@@ -933,6 +179,7 @@ void SemanticAnalysisVisitor::visit(ast::FunctionDefinition& function) {
     function.setArguments(symbolTable.getCurrentScopeArguments());
     function.setLocalVariables(symbolTable.getCurrentScopeSymbols());
     symbolTable.endFunction();
+    currentReturnType.reset();
 }
 
 void SemanticAnalysisVisitor::visit(ast::Block& block) {
@@ -944,22 +191,21 @@ void SemanticAnalysisVisitor::visit(ast::Block& block) {
 void SemanticAnalysisVisitor::typeCheck(const type::Type& typeFrom, const type::Type& typeTo,
         const translation_unit::Context& context)
 {
-    if (!typeTo.canAssignFrom(typeFrom)) {
-        semanticError("type mismatch: can't convert " + typeFrom.to_string() + " to " + typeTo.to_string(), context);
+    if (typeTo.canAssignFrom(typeFrom)) {
+        return;
     }
+    semanticError(type::productAssignFailureMessage(typeTo, typeFrom), context);
 }
 
 void SemanticAnalysisVisitor::rejectFunctionValue(const type::Type& type, const translation_unit::Context& context) {
-    // Pointer-to-function types are stored as function types with indirection > 0, so they
-    // also report isFunction(). Only bare function designators (no pointer) are rejected.
-    if (type.isFunction() && !type.isPointer()) {
+    if (type::isBareFunction(type)) {
         semanticError("function designator used as a value is not supported", context);
     }
 }
 
 void SemanticAnalysisVisitor::semanticError(std::string message, const translation_unit::Context& context) {
     containsSemanticErrors = true;
-    err << context << ": error: " << message << "\n";
+    semanticErrorLogger() << context << ": error: " << message << "\n";
 }
 
 bool SemanticAnalysisVisitor::successfulSemanticAnalysis() const {
@@ -974,5 +220,5 @@ std::vector<ValueEntry> SemanticAnalysisVisitor::getGlobalVariables() const {
     return symbolTable.getGlobalVariables();
 }
 
-} // namespace semantic_analyzer
 
+} // namespace semantic_analyzer
