+++ src/symbols/CMakeLists.txt
@@ -1,0 +1,7 @@
+add_library(symbols
+        AnnotationStore.cpp
+        AnnotationStore.h
+        AddressPlan.h
+        )
+
+trans_target_defaults(symbols)
